@@ -26,12 +26,11 @@
     },
     "devDependencies": {
         "@dhis2/cli-app-scripts": "12.0.0-alpha.13",
-        "@testing-library/react": "^12.0.0",
-        "@testing-library/react-hooks": "^8.0.1",
+        "@testing-library/react": "^16.0.1",
         "enzyme": "^3.11.0",
         "enzyme-adapter-react-16": "^1.15.5",
-        "react": "^16.8",
-        "react-dom": "^16.8"
+        "react": "^18",
+        "react-dom": "^18"
     },
     "scripts": {
         "build": "d2-app-scripts build",
@@ -44,8 +43,8 @@
         "classnames": "^2",
         "moment": "^2",
         "prop-types": "^15",
-        "react": "^16.8",
-        "react-dom": "^16.8",
+        "react": "^18",
+        "react-dom": "^18",
         "styled-jsx": "^4"
     },
     "jest": {
