@@ -24,8 +24,8 @@
         "moment": "^2.29.1",
         "post-robot": "^10.0.46",
         "prop-types": "^15.7.2",
-        "react": "^16.8.6",
-        "react-dom": "^16.8.6",
+        "react": "^18",
+        "react-dom": "^18",
         "source-map-explorer": "^2.1.0",
         "styled-jsx": "^4.0.1",
         "typeface-roboto": "^0.0.75",
